requirements.txt
# ML-Driven School Schedule Generation System Requirements

# Core Data Processing
pandas>=2.2.3,<3
numpy>=2.1.1,<3

# Machine Learning
scikit-learn>=1.5.2,<2
xgboost>=2.0.0,<3

# Configuration Management
PyYAML>=6.0.2,<7

# Data Export
openpyxl>=3.1.2,<4
XlsxWriter>=3.1.0,<4

# Utilities
python-dateutil>=2.8.2,<3

# Optional: Advanced ML (uncomment if needed)
# catboost>=1.2.0
# lightgbm>=4.0.0
# torch>=2.0.0
# tensorflow>=2.13.0

# Optional: Visualization (uncomment if needed)
# matplotlib>=3.9.2,<4
# seaborn>=0.13.2,<1
# plotly>=5.15.0

# Optional: Graph Processing (uncomment if needed)
# networkx>=3.0,<4
